fix(color) overflow with 16 bit color depth

+++ src/misc/lv_color.c
@@ -43,7 +43,7 @@ LV_ATTRIBUTE_FAST_MEM void lv_color_fill(lv_color_t * buf, lv_color_t color, uin
         px_num--;
     }
 
-    uint32_t c32 = color.full + (color.full << 16);
+    uint32_t c32 = (uint32_t)color.full + ((uint32_t)color.full << 16);
     uint32_t * buf32 = (uint32_t *)buf;
 
     while(px_num > 16) {
